fix: updated story to match mdx
