+++ modules/react/avatar/README.md
@@ -8,25 +8,6 @@ A component showing a user's photo with a circular crop.
 yarn add @workday/canvas-kit-react
 ```
 
-## Usage
-
-```tsx
-import * as React from 'react';
-import {Avatar, AvatarVariant} from '@workday/canvas-kit-react/avatar';
-
-// Basic button with click handler
-<Avatar onClick={someFunction} />
-
-// Rendering as a div instead of a button
-<Avatar as="div" />
-
-// Using static properties on Avatar
-<Avatar
-  size="extraSmall"
-  variant="dark"
-/>
-```
-
 ## Component Props
 
 ### Required
@@ -46,10 +27,6 @@ Default: "light"
 | "light" | Light grey background, dark icon |
 | "dark"  | Dark blue background, white icon |
 
-```tsx
-<Avatar variant="dark" />
-```
-
 ---
 
 #### `size: "extraSmall" | "small" | "medium" | "large" | "extraLarge" | "extraExtraLarge" | (string & {})`
@@ -67,25 +44,11 @@ Default: "medium"
 | "extraLarge"      | 64        |
 | "extraExtraLarge" | 120       |
 
-```tsx
-<Avatar size="large" />
-
-// Custom Size.
-// Note: Please use the predefined sizes as it works nicely with the Canvas Design
-System
-
-<Avatar size="48px"/>
-```
-
 ---
 
 #### `altText: string`
 
 > The alt text of the Avatar image. This prop is also used for the aria-label.
-
-```tsx
-<Avatar altText="User-Avatar-Button" />
-```
 
 ---
 
@@ -93,19 +56,11 @@ System
 
 > The URL of the user's photo. Expects a square (1:1) photo.
 
-```tsx
-<Avatar url="https://example.com/image.png" />
-```
-
 ---
 
 #### `onClick: (e: React.MouseEvent<HTMLButtonElement>) => void`
 
 > An onClick function to pass to the underlying `<button>` element.
-
-```tsx
-<Avatar onClick={fn} />
-```
 
 ---
 
@@ -113,26 +68,14 @@ System
 
 > An objectFit property that can customize how to resize your image to fit its container.
 
-```tsx
-<Avatar url="https://example.com/image.png" objectFit="cover" />
-```
-
 ---
 
 #### `ref: React.Ref<HTMLButtonElement>`
 
 > A ref to the underlying `<button>` or `<div>` element.
 
-```tsx
-<Avatar ref={ref} />
-```
-
 ---
 
 #### `as: 'div'`
 
 > Option to display the avatar as a div instead of a button.
-
-```tsx
-<Avatar as="div" />
-```
