+++ modules/react/avatar/stories/avatar/Avatar.stories.mdx
@@ -1,0 +1,103 @@
+import {Specifications, SymbolDoc, SymbolDescription} from '@workday/canvas-kit-docs';
+
+import {DarkVariantDiv} from './examples/div/DarkVariantDiv';
+import {DynamicBackgroundColorDiv} from './examples/div/DynamicBackgroundColorDiv';
+import {DynamicSizeDiv} from './examples/div/DynamicSizeDiv';
+import {ImageDiv} from './examples/div/ImageDiv';
+import {LazyLoadDiv} from './examples/div/LazyLoadDiv';
+import {LightVariantDiv} from './examples/div/LightVariantDiv';
+import {ObjectFitDiv} from './examples/div/ObjectFitDiv';
+
+import {DarkVariantButton} from './examples/button/DarkVariantButton';
+import {ImageButton} from './examples/button/ImageButton.tsx';
+import {LightVariantButton} from './examples/button/LightVariantButton';
+
+<Meta title="Components/Avatar" />
+
+# Canvas Kit Avatar
+
+A component showing a user's photo with a circular crop.
+
+[> Workday Design Reference](https://canvas.workdaydesign.com/components/indicators/avatar)
+
+## Installation
+
+```sh
+yarn add @workday/canvas-kit-react
+```
+
+## Usage
+
+<SymbolDescription name="Avatar" fileName="/react/" />
+
+The example below shows multiple instances of a `Avatar` with various configurations. Like many of
+our components, Avatar accepts an `as` prop, which lets you change the underlying semantic element
+to a `<div>`. This should be done with caution to ensure the best accessibility. Generally,
+`<button>` elements should be used for actions.
+
+### Variant and Size
+
+Avatar defaults to using a "light" `variant` property. You can change the `variant` prop to "dark"
+by including `variant="dark"`.
+
+Avatar defaults to using a "medium" `size` property. You can change the `size` prop to various
+sizing options. See Component API for more details.
+
+#### As a `<div>`
+
+<ExampleCodeBlock code={LightVariantDiv} />
+
+<ExampleCodeBlock code={DarkVariantDiv} />
+
+#### As a `<button>`
+
+<ExampleCodeBlock code={LightVariantButton} />
+
+<ExampleCodeBlock code={DarkVariantButton} />
+
+### Image
+
+Avatar will convert to an image avatar when the `url` prop is included with acceptable url.
+
+#### As a `<div>`
+
+<ExampleCodeBlock code={ImageDiv} />
+
+#### As a `<button>`
+
+<ExampleCodeBlock code={ImageButton} />
+
+### Object Fit
+
+Avatar defaults to using `contain` for object-fit. You can change this property to any other
+acceptable object-fit properties.
+
+<ExampleCodeBlock code={ObjectFitDiv} />
+
+### Lazy Load
+
+Example of implementing lazy load behavior with Avatar.
+
+<ExampleCodeBlock code={LazyLoadDiv} />
+
+### Dynamic Behavior
+
+#### Dynamic Size
+
+For the `size` prop, you can explicitly specify the size you want for the Avatar.
+
+<ExampleCodeBlock code={DynamicSizeDiv} />
+
+#### Dynamic Color
+
+To change the background color of the Avatar, you can use our createStencil functionality.
+
+<ExampleCodeBlock code={DynamicBackgroundColorDiv} />
+
+## Component API
+
+<SymbolDoc name="Avatar" fileName="/react/" />
+
+## Specifications
+
+<Specifications file="Avatar.spec.ts" name="Avatar" />
